// test/src/model/evil/baseEvil.ts
import { GenerateView, IgnoreViewModel } from "../../../..";
import {NotExistModelNodeModules} from 'node-modules-model';
import { Hero } from "../hero/hero";

@GenerateView({
    model: "BaseEvilViewModel",
    type: 'class',
    filePath: "./test/dist/_autogenerated_viewmodels/evils",
    mapperPath: "./test/dist/_autogenerated_viewmodels/evils/mappers"})
export class BaseEvil {
    id: number;

    name: string;

    nodeModulesModel: NotExistModelNodeModules;

    lastKilledHero?: Hero;

    @IgnoreViewModel()
    ignoringField: string;

    @IgnoreViewModel()
    notIgnoringFieldInDoctorEvil: boolean;
}